Add files via upload
I wrote the main part of the app, but I have a real problem with connecting with a Spotify API. I can't get a token from Spotify. My authorization runs successfully, but when I insert the response URL to terminal to get token - a mistake happens. So the app could fail because I can't get that token.